+++ source/screens/index.js
@@ -4,5 +4,5 @@ export { default as LoginScreen } from "./LoginScreen";
 export { default as ProfileScreen } from "./ProfileScreen";
 export { default as SplashScreen } from "./SplashScreen";
 export { default as AboutScreen } from "./AboutScreen";
-export { default as BookingFormScreen } from "./BookingFormScreen";
+export { default as BookingFormScreen } from "./bookingScreens/BookingFormScreen";
 export { default as BookingSummary } from "./bookingScreens/BookingSummary";
